# docs/openapi.yaml
openapi: 3.1.0
info:
  title: adaptcard API
  version: 0.2.0
  description: |
    Expanded OpenAPI contract covering the current integration-tested adaptcard API surface.
servers:
  - url: http://127.0.0.1:8787
paths:
...
  /notes:
    post:
      summary: Create note with template-based card generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deckId, front, back]
              properties:
                deckId: { type: string }
                knowledgePointId: { type: string }
                front: { type: string }
                back: { type: string }
                tags:
                  type: array
                  items: { type: string }
                cardType:
                  type: string
                  enum: [basic, reverse, cloze]
                  default: basic
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  cardId: { type: string }
                  cardIds:
                    type: array
                    items: { type: string }
                  cardCount: { type: integer }
    get:
      summary: List notes
      parameters:
        - in: query
          name: deckId
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Note list
          content:
            application/json:
              schema:
                type: object
  /notes/{id}/review-history:
    get:
      summary: List review history timeline for one note (linked knowledge point)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Note review history page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteReviewHistoryResponse'
...
components:
  schemas:
...
    KnowledgePointReviewHistoryItem:
      type: object
      properties:
        id: { type: string }
        sessionId:
          type: string
          nullable: true
        cardId:
          type: string
          nullable: true
        reviewedAt: { type: string, format: date-time }
        rating: { type: integer }
        correctRate: { type: number }
        stats:
          type: object
          properties:
            total:
              type: integer
              nullable: true
            correct:
              type: integer
              nullable: true
        answers:
          type: array
          items:
            type: object
            properties:
              questionId: { type: string }
              userAnswer: { type: string }
...
    NoteReviewHistoryResponse:
      type: object
      properties:
        noteId: { type: string }
        knowledgePointId:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgePointReviewHistoryItem'
        page:
          type: object
          properties:
            limit: { type: integer }
            offset: { type: integer }
            total: { type: integer }
...